+++ io_new.h
@@ -30,7 +30,7 @@ void IO_Init(void);
 void IO_Destroy(void);
 uint16_t IO_Read(uint32_t address);
 void IO_Write(uint32_t address, uint16_t value);
-uint16_t IO_Ident(uint16_t level);
+int IO_Ident(uint16_t level);
 void IO_Tick(void);
 
 extern void device_interrupt(ushort interruptBits);
